Create README.md
# AI All-in-One App

This is an application that aims to be an advanced "all-knowing" assistant. It will integrate with AI, books, and other resources to provide solutions and guidance to users in various aspects of life.

## How to Contribute

I'm looking for help with:
- Designing the user interface
- Writing code (mainly in Python)
- Testing the app for bugs
- Suggestions for additional features

If you're interested in contributing, please fork the repository and send a pull request.